Pd files read/write from repo directory; separate get and set filenames
Two changes:

- I first wrote and tested the Pd (pure data) files in a directory in my
  home directory. I have changed them to consider this project repo
their home.

- When reading a preset from the device, it creates a file with a name
  like "lpd8-mk2-get-preset-1.txt"; when writing a preset, it looks for
"lpd8-mk2-set-preset-1.txt". One reason for doing this: when reading, we
read the sysex "preamble" and "postamble" into the file; when writing,
the Pd code generates the preamble and postamble, and the file should
contain just the settings payload (164 bytes - which is to say, 164
lines, with a single integere on each line).