Generate presets with sysex header and footer; use correct length
Previously the Lua code was generating just the "guts" of the preset (as
a string of numbers) and letting the "plumbing" (eg, puredata) frame it
as a sysex.

I think it's better to generate full sysexes, and the plumbing can do
whatever it wants. In this case, the full sysex is for preset 0 - the
temporary one - so if something goes wrong, no presets are harmed.

The puredata code in fact *ignores* the sysex framing in the file and
creates its own, including choosing a proper preset number. Doing things
this way seems most simple and flexible.

As an experiment I tried using 165 - the correct number - as the payload
size, instead of the 169 that the device uses when returning a preset.
It seems to work! So, either one works. I'd rather use the "correct"
one, as long as it doesn't break the device!